move handle keys context changing to configuration so that it is updated when you update config without needing to restart
